test: stabilize page size pagination test
