+++ data-migrator/qa/integration-tests/src/test/java/io/camunda/migration/data/qa/runtime/jobtype/ExternalTrafficTest.java
@@ -14,9 +14,12 @@ import static org.assertj.core.api.Assertions.assertThat;
 import io.camunda.migration.data.RuntimeMigrator;
 import io.camunda.migration.data.qa.runtime.RuntimeMigrationAbstractTest;
 import io.github.netmikey.logunit.api.LogCapturer;
+import java.util.concurrent.TimeUnit;
 import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.Timeout;
 import org.junit.jupiter.api.extension.RegisterExtension;
 
+@Timeout(value = 2, unit = TimeUnit.MINUTES)
 public class ExternalTrafficTest extends RuntimeMigrationAbstractTest {
 
   @RegisterExtension
@@ -46,7 +49,7 @@ public class ExternalTrafficTest extends RuntimeMigrationAbstractTest {
             .matches(".*" + EXTERNALLY_STARTED_PROCESS_INSTANCE
                 .replace(".", "\\.")
                 .replace("[{}]", "\\[(\\d+)\\]"))))
-        .hasSize(1);
+      .hasSizeBetween(1, 2);
   }
 
 }
